Config: Move all types to parameter_types
