+++ tests/output/test_parameters_keytype.py
@@ -1,36 +1,32 @@
 from firedrake.parameters import Parameters
+from firedrake.parameter_types import *
 
 
 def test_inttype_inference():
-    from firedrake.parameters import IntType
     params = Parameters()
     params["int"] = 1
     assert isinstance(params.get_key("int").type, IntType)
 
 
 def test_floattype_inference():
-    from firedrake.parameters import FloatType
     params = Parameters()
     params["float"] = 1.1
     assert isinstance(params.get_key("float").type, FloatType)
 
 
 def test_strtype_inference():
-    from firedrake.parameters import StrType
     params = Parameters()
     params["str"] = "str"
     assert isinstance(params.get_key("str").type, StrType)
 
 
 def test_booltype_inference():
-    from firedrake.parameters import BoolType
     params = Parameters()
     params["bool"] = True
     assert isinstance(params.get_key("bool").type, BoolType)
 
 
 def test_int_validate_ints():
-    from firedrake.parameters import IntType
     int_type = IntType()
     assert int_type.validate(1)
     assert int_type.validate(0)
@@ -38,7 +34,6 @@ def test_int_validate_ints():
 
 
 def test_int_validate_bounded_both_sides():
-    from firedrake.parameters import IntType
     int_type = IntType(lower_bound=1, upper_bound=3)
     assert not int_type.validate(0)
     assert int_type.validate(1)
@@ -48,7 +43,6 @@ def test_int_validate_bounded_both_sides():
 
 
 def test_int_validate_bounded_upper():
-    from firedrake.parameters import IntType
     int_type = IntType(upper_bound=3)
     assert int_type.validate(0)
     assert int_type.validate(1)
@@ -58,7 +52,6 @@ def test_int_validate_bounded_upper():
 
 
 def test_int_validate_bounded_lower():
-    from firedrake.parameters import IntType
     int_type = IntType(lower_bound=1)
     assert not int_type.validate(0)
     assert int_type.validate(1)
@@ -68,14 +61,12 @@ def test_int_validate_bounded_lower():
 
 
 def test_int_validate_floats():
-    from firedrake.parameters import IntType
     int_type = IntType()
     assert not int_type.validate(1.0)
     assert not int_type.validate(1.5)
 
 
 def test_int_validate_strs():
-    from firedrake.parameters import IntType
     int_type = IntType()
     assert int_type.validate("1")
     assert int_type.validate("0")
@@ -85,7 +76,6 @@ def test_int_validate_strs():
 
 
 def test_int_parse_valid():
-    from firedrake.parameters import IntType
     int_type = IntType()
     assert int_type.parse("1") == 1
     assert int_type.parse("0") == 0
@@ -93,14 +83,12 @@ def test_int_parse_valid():
 
 
 def test_int_parse_unbounded_invalid():
-    from firedrake.parameters import IntType
     int_type = IntType()
     assert int_type.parse("1.0") is None
     assert int_type.parse("1.5") is None
 
 
 def test_int_parse_bounded_invalid():
-    from firedrake.parameters import IntType
     int_type = IntType(lower_bound=1, upper_bound=3)
     assert int_type.parse("4") is None
     assert int_type.parse("0") is None
@@ -108,7 +96,6 @@ def test_int_parse_bounded_invalid():
 
 def test_float_validate_ints():
     # Float type accepts ints
-    from firedrake.parameters import FloatType
     float_type = FloatType()
     assert float_type.validate(1)
     assert float_type.validate(0)
@@ -116,7 +103,6 @@ def test_float_validate_ints():
 
 
 def test_float_validate_bounded_both_sides():
-    from firedrake.parameters import FloatType
     float_type = FloatType(lower_bound=1, upper_bound=3)
     assert not float_type.validate(0.99)
     assert float_type.validate(1.0)
@@ -129,7 +115,6 @@ def test_float_validate_bounded_both_sides():
 
 
 def test_float_validate_bounded_upper():
-    from firedrake.parameters import FloatType
     float_type = FloatType(upper_bound=3)
     assert float_type.validate(0.0)
     assert float_type.validate(1.0)
@@ -142,7 +127,6 @@ def test_float_validate_bounded_upper():
 
 
 def test_float_validate_bounded_lower():
-    from firedrake.parameters import FloatType
     float_type = FloatType(lower_bound=1)
     assert not float_type.validate(0.99)
     assert float_type.validate(1.0)
@@ -155,7 +139,6 @@ def test_float_validate_bounded_lower():
 
 
 def test_float_validate_floats():
-    from firedrake.parameters import FloatType
     float_type = FloatType()
     assert float_type.validate(1.0)
     assert float_type.validate(1.5)
@@ -165,7 +148,6 @@ def test_float_validate_floats():
 
 
 def test_float_validate_strs():
-    from firedrake.parameters import FloatType
     float_type = FloatType()
     assert float_type.validate("1")
     assert float_type.validate("0")
@@ -178,7 +160,6 @@ def test_float_validate_strs():
 
 
 def test_float_parse_valid():
-    from firedrake.parameters import FloatType
     from math import isnan
     float_type = FloatType()
     assert float_type.parse("1.0") == 1.0
@@ -193,7 +174,6 @@ def test_float_parse_valid():
 
 
 def test_float_parse_bounded_invalid():
-    from firedrake.parameters import FloatType
     float_type = FloatType(lower_bound=1, upper_bound=3)
     assert float_type.parse("4.0") is None
     assert float_type.parse("0.0") is None
@@ -203,21 +183,18 @@ def test_float_parse_bounded_invalid():
 
 
 def test_bool_validate_bools():
-    from firedrake.parameters import BoolType
     bool_type = BoolType()
     assert bool_type.validate(True)
     assert bool_type.validate(False)
 
 
 def test_bool_validate_strs():
-    from firedrake.parameters import BoolType
     bool_type = BoolType()
     assert bool_type.validate("True")
     assert bool_type.validate("False")
 
 
 def test_bool_validate_other():
-    from firedrake.parameters import BoolType
     bool_type = BoolType()
     assert not bool_type.validate(1)
     assert not bool_type.validate(0)
@@ -232,7 +209,6 @@ def test_bool_validate_other():
 
 
 def test_bool_parse_valid():
-    from firedrake.parameters import BoolType
     bool_type = BoolType()
     assert bool_type.parse(True)
     assert not bool_type.parse(False)
@@ -241,7 +217,6 @@ def test_bool_parse_valid():
 
 
 def test_bool_parse_invalid():
-    from firedrake.parameters import BoolType
     bool_type = BoolType()
     assert bool_type.parse(1) is None
     assert bool_type.parse(0) is None
@@ -256,13 +231,11 @@ def test_bool_parse_invalid():
 
 
 def test_str_no_options():
-    from firedrake.parameters import StrType
     str_type = StrType()
     assert str_type.options == []
 
 
 def test_str_with_options():
-    from firedrake.parameters import StrType
     str_type = StrType("A", "B", "C", "D")
     assert "A" in str_type.options
     assert "B" in str_type.options
@@ -272,7 +245,6 @@ def test_str_with_options():
 
 
 def test_str_clear_options():
-    from firedrake.parameters import StrType
     str_type = StrType("A", "B", "C", "D")
     assert "A" in str_type.options
     assert "B" in str_type.options
@@ -288,7 +260,6 @@ def test_str_clear_options():
 
 
 def test_str_add_options():
-    from firedrake.parameters import StrType
     str_type = StrType("A", "B", "C", "D")
     assert "A" in str_type.options
     assert "B" in str_type.options
@@ -304,7 +275,6 @@ def test_str_add_options():
 
 
 def test_str_validate_str_with_validate_function():
-    from firedrake.parameters import StrType
     str_type = StrType()
     str_type.set_validate_function(lambda x: len(x) == 1)
     assert str_type.validate("a")
@@ -317,7 +287,6 @@ def test_str_validate_str_with_validate_function():
 
 
 def test_str_validate_str_no_options():
-    from firedrake.parameters import StrType
     str_type = StrType()
     assert str_type.validate("1")
     assert str_type.validate("random stuff")
@@ -326,7 +295,6 @@ def test_str_validate_str_no_options():
 
 
 def test_str_validate_str_with_options():
-    from firedrake.parameters import StrType
     str_type = StrType("A", "B", "")
     assert str_type.validate("A")
     assert str_type.validate("B")
@@ -335,7 +303,6 @@ def test_str_validate_str_with_options():
 
 
 def test_str_validate_others():
-    from firedrake.parameters import StrType
     str_type = StrType()
     assert not str_type.validate(None)
     assert not str_type.validate(1)
@@ -346,7 +313,6 @@ def test_str_validate_others():
 
 
 def test_str_parse_valid():
-    from firedrake.parameters import StrType
     str_type = StrType()
     assert str_type.parse("1") == "1"
     assert str_type.parse("random stuff") == "random stuff"
@@ -355,7 +321,6 @@ def test_str_parse_valid():
 
 
 def test_str_parse_invalid():
-    from firedrake.parameters import StrType
     str_type = StrType()
     assert str_type.parse(None) is None
     assert str_type.parse(1) is None
@@ -366,7 +331,6 @@ def test_str_parse_invalid():
 
 
 def test_str_validate_unicode():
-    from firedrake.parameters import StrType
     str_type = StrType()
     assert str_type.validate(u"1")
     assert str_type.validate(u"random stuff")
@@ -375,7 +339,6 @@ def test_str_validate_unicode():
 
 
 def test_str_parse_unicode():
-    from firedrake.parameters import StrType
     str_type = StrType()
     assert str_type.parse(u"1") == "1"
     assert str_type.parse(u"random stuff") == "random stuff"
@@ -384,7 +347,6 @@ def test_str_parse_unicode():
 
 
 def test_or_type_validate():
-    from firedrake.parameters import OrType, StrType, IntType
     or_type = OrType(StrType("This"), IntType(1, 1))
     assert or_type.validate("This")
     assert or_type.validate(1)
@@ -397,7 +359,6 @@ def test_or_type_validate():
 
 
 def test_or_type_parse():
-    from firedrake.parameters import OrType, StrType, IntType
     or_type = OrType(StrType("This"), IntType(1, 1))
     assert or_type.parse("This") == "This"
     assert or_type.parse(1) == 1
